+++ CHANGELOG.md
@@ -6,6 +6,9 @@ https://github.com/clojure/brew-install/blob/1.10.1/CHANGELOG.md
 
 ## Releases
 
+- Fifty-second release [2022-12-15]
+    - Based on clojure tools v.1.11.0.1208
+
 - Fifty-first release [2022-11-15]
     - Based on clojure tools v.1.11.0.1200